Add example for writeMicroseconds()
The writeMicroseconds() function is used to mimic the Arduino Servo library writeMicroseconds() behavior.
The writeMicroseconds() function is based on setServoPulse() but uses microseconds instead of seconds. This is to match with the Arduino Servo library writeMicroseconds() behavior.
This example follows what setPWM() example does driving the servo through the range of positions but using writeMicroseconds()
writeMicroseconds() is not expected to be precise due to calculation rounding converting between microseconds and ticks.
New constants were added to simplify changes like the frequency of the PCA9685 chip.