Update int.osc. notes in examples
In theory the data sheet says the internal oscillator (clock) is 25MHz but in testing the internal oscillator for the PCA9685 chip is a range between about 23-27MHz.
Adafruit testing for implimenting the int.osc. showed and upper level of about 27 Mhz
My own testing with 4 different Adafruit PCA9685 16-Channel Servo Drivers gave 24060000, 24090000, 24600000, and 25000000 Mhz when tuning the int.osc. frequency to get aproximatly 50Hz PWM signals. Actual pwm signals were 50.6xx to 50.1xx Hz across the four chips tested as the  PCA9685 chip output is not super precise.

This PR expands the notes on the PWM frequency and gives rough instructions on checking the PWM frequenty for tuning the int.osc. value.